+++ android/expoview/src/main/java/versioned/host/exp/exponent/ExperiencePackagePicker.kt
@@ -17,7 +17,7 @@ import expo.modules.clipboard.ClipboardModule
 import expo.modules.constants.ConstantsPackage
 import expo.modules.contacts.ContactsPackage
 import expo.modules.core.interfaces.Package
-import expo.modules.crypto.CryptoPackage
+import expo.modules.crypto.CryptoModule
 import expo.modules.device.DevicePackage
 import expo.modules.documentpicker.DocumentPickerPackage
 import expo.modules.easclient.EASClientModule
@@ -32,7 +32,7 @@ import expo.modules.gl.GLPackage
 import expo.modules.haptics.HapticsPackage
 import expo.modules.imageloader.ImageLoaderPackage
 import expo.modules.imagemanipulator.ImageManipulatorPackage
-import expo.modules.imagepicker.ImagePickerPackage
+import expo.modules.imagepicker.ImagePickerModule
 import expo.modules.intentlauncher.IntentLauncherPackage
 import expo.modules.keepawake.KeepAwakePackage
 import expo.modules.kotlin.ModulesProvider
@@ -49,6 +49,7 @@ import expo.modules.network.NetworkPackage
 import expo.modules.notifications.NotificationsPackage
 import expo.modules.permissions.PermissionsPackage
 import expo.modules.print.PrintPackage
+import expo.modules.random.RandomModule
 import expo.modules.screencapture.ScreenCapturePackage
 import expo.modules.screenorientation.ScreenOrientationPackage
 import expo.modules.securestore.SecureStorePackage
@@ -80,7 +81,6 @@ object ExperiencePackagePicker : ModulesProvider {
     CameraPackage(),
     ConstantsPackage(),
     ContactsPackage(),
-    CryptoPackage(),
     DevicePackage(),
     DocumentPickerPackage(),
     ErrorRecoveryPackage(),
@@ -94,7 +94,6 @@ object ExperiencePackagePicker : ModulesProvider {
     HapticsPackage(),
     ImageLoaderPackage(),
     ImageManipulatorPackage(),
-    ImagePickerPackage(),
     IntentLauncherPackage(),
     KeepAwakePackage(),
     LocalAuthenticationPackage(),
@@ -142,8 +141,11 @@ object ExperiencePackagePicker : ModulesProvider {
   override fun getModulesList(): List<Class<out Module>> = listOf(
     CellularModule::class.java,
     ClipboardModule::class.java,
+    CryptoModule::class.java,
     EASClientModule::class.java,
+    ImagePickerModule::class.java,
     LinearGradientModule::class.java,
+    RandomModule::class.java,
     WebBrowserModule::class.java,
   )
 }