+++ docs/components/plugins/permissions/AndroidPermissions.tsx
@@ -6,6 +6,7 @@ import { androidPermissions, AndroidPermission, PermissionReference } from './da
 import Permalink from '~/components/Permalink';
 import { InlineCode } from '~/components/base/code';
 import { Quote } from '~/components/base/paragraph';
+import { Cell, HeaderCell, Row, Table, TableHead } from '~/ui/components/Table';
 
 // TODO(cedric): all commented code is related to the "granter" column.
 // This column defines if the permission is granted by the system or user (requires notification).
@@ -19,47 +20,51 @@ type AndroidPermissionsProps = {
 
 // const grantedByInfo = 'Some permissions are granted by the system without user approval';
 
-export function AndroidPermissions(props: AndroidPermissionsProps) {
-  const list = React.useMemo(() => getPermissions(props.permissions), [props.permissions]);
+export function AndroidPermissions({ permissions }: AndroidPermissionsProps) {
+  const list = React.useMemo(() => getPermissions(permissions), [permissions]);
 
   return (
-    <table>
-      <thead>
-        <tr>
-          <th>Android Permission</th>
-          {/* <th>
+    <Table>
+      <TableHead>
+        <Row>
+          <HeaderCell>Android Permission</HeaderCell>
+          {/* <HeaderCell>
             <span css={grantedByInfoStyle} title={grantedByInfo}>
               Granted by <QuestionIcon size={12} title={grantedByInfo} />
             </span>
-          </th> */}
-          <th>Description</th>
-        </tr>
-      </thead>
+          </HeaderCell> */}
+          <HeaderCell>Description</HeaderCell>
+        </Row>
+      </TableHead>
       <tbody>
         {list.map(permission => (
           <AndroidPermissionRow key={permission.name} {...permission} />
         ))}
       </tbody>
-    </table>
+    </Table>
   );
 }
 
-function AndroidPermissionRow(permission: AndroidPermission) {
-  const { name, description, explanation, warning, apiDeprecated } = permission;
-
+function AndroidPermissionRow({
+  name,
+  description,
+  explanation,
+  warning,
+  apiDeprecated,
+}: AndroidPermission) {
   return (
-    <tr css={apiDeprecated && deprecatedStyle}>
-      <td>
+    <Row subtle={!!apiDeprecated}>
+      <Cell>
         <Permalink id={`permission-${name.toLowerCase()}`}>
           <span>
             <InlineCode>{name}</InlineCode>
           </span>
         </Permalink>
-      </td>
-      {/* <td>
+      </Cell>
+      {/* <Cell>
         <i>{getPermissionGranter(permission)}</i>
-      </td> */}
-      <td>
+      </Cell> */}
+      <Cell>
         {!!description && (
           <p css={(warning || explanation) && descriptionSpaceStyle}>{description}</p>
         )}
@@ -73,8 +78,8 @@ function AndroidPermissionRow(permission: AndroidPermission) {
             <span dangerouslySetInnerHTML={{ __html: explanation }} />
           </Quote>
         )}
-      </td>
-    </tr>
+      </Cell>
+    </Row>
   );
 }
 
@@ -91,10 +96,6 @@ function getPermissions(permissions: AndroidPermissionsProps['permissions']) {
 // const grantedByInfoStyle = css`
 //   white-space: nowrap;
 // `;
-
-const deprecatedStyle = css`
-  opacity: 0.5;
-`;
 
 const descriptionSpaceStyle = css`
   margin-bottom: 1rem;