+++ packages/expo-modules-core/android/src/main/cpp/JSIInteropModuleRegistry.cpp
@@ -20,7 +20,12 @@ JSIInteropModuleRegistry::initHybrid(jni::alias_ref<jhybridobject> jThis) {
 void JSIInteropModuleRegistry::registerNatives() {
   registerHybrid({
                    makeNativeMethod("initHybrid", JSIInteropModuleRegistry::initHybrid),
-                   makeNativeMethod("installJSI", JSIInteropModuleRegistry::installJSI)
+                   makeNativeMethod("installJSI", JSIInteropModuleRegistry::installJSI),
+                   makeNativeMethod("installJSIForTests",
+                                    JSIInteropModuleRegistry::installJSIForTests),
+                   makeNativeMethod("evaluateScript", JSIInteropModuleRegistry::evaluateScript),
+                   makeNativeMethod("global", JSIInteropModuleRegistry::global),
+                   makeNativeMethod("createObject", JSIInteropModuleRegistry::createObject),
                  });
 }
 
@@ -33,9 +38,11 @@ void JSIInteropModuleRegistry::installJSI(
   jni::alias_ref<react::CallInvokerHolder::javaobject> nativeInvokerHolder
 ) {
   auto runtime = reinterpret_cast<jsi::Runtime *>(jsRuntimePointer);
-  runtimeHolder = std::make_unique<JavaScriptRuntime>(runtime);
-  jsInvoker = jsInvokerHolder->cthis()->getCallInvoker();
-  nativeInvoker = nativeInvokerHolder->cthis()->getCallInvoker();
+  runtimeHolder = std::make_shared<JavaScriptRuntime>(
+    runtime,
+    jsInvokerHolder->cthis()->getCallInvoker(),
+    nativeInvokerHolder->cthis()->getCallInvoker()
+  );
 
   auto expoModules = std::make_shared<ExpoModulesHostObject>(this);
   auto expoModulesObject = jsi::Object::createFromHostObject(*runtime, expoModules);
@@ -44,6 +51,22 @@ void JSIInteropModuleRegistry::installJSI(
     ->global()
     .setProperty(
       *runtime,
+      "ExpoModules",
+      std::move(expoModulesObject)
+    );
+}
+
+void JSIInteropModuleRegistry::installJSIForTests() {
+  runtimeHolder = std::make_shared<JavaScriptRuntime>();
+  jsi::Runtime &jsiRuntime = *runtimeHolder->get();
+
+  auto expoModules = std::make_shared<ExpoModulesHostObject>(this);
+  auto expoModulesObject = jsi::Object::createFromHostObject(jsiRuntime, expoModules);
+
+  jsiRuntime
+    .global()
+    .setProperty(
+      jsiRuntime,
       "ExpoModules",
       std::move(expoModulesObject)
     );
@@ -63,5 +86,19 @@ JSIInteropModuleRegistry::callGetJavaScriptModuleObjectMethod(const std::string 
 jni::local_ref<JavaScriptModuleObject::javaobject>
 JSIInteropModuleRegistry::getModule(const std::string &moduleName) const {
   return callGetJavaScriptModuleObjectMethod(moduleName);
+}
+
+jni::local_ref<JavaScriptValue::javaobject> JSIInteropModuleRegistry::evaluateScript(
+  jni::JString script
+) {
+  return runtimeHolder->evaluateScript(script.toStdString());
+}
+
+jni::local_ref<JavaScriptObject::javaobject> JSIInteropModuleRegistry::global() {
+  return runtimeHolder->global();
+}
+
+jni::local_ref<JavaScriptObject::javaobject> JSIInteropModuleRegistry::createObject() {
+  return runtimeHolder->createObject();
 }
 } // namespace expo