+++ packages/expo-dev-menu/CHANGELOG.md
@@ -6,7 +6,11 @@
 
 ### 🎉 New features
 
+- Add JS API to register buttons in dev menu. ([#17528](https://github.com/expo/expo/pull/17528) by [@ajsmth](https://github.com/ajsmth))
+
 ### 🐛 Bug fixes
+
+- Update dev settings when initial props change. ([#17663](https://github.com/expo/expo/pull/17663) by [@esamelson](https://github.com/esamelson))
 
 ### 💡 Others